Removed delay() call from Bluetooth.readPacket().
Previously, I needed a mysterious call to delay() in order for the readPacket()
method to work correctly. Without the delay, the program would sporadically
hang.

I believe that the hang occurred whenever readPacket() was called when the
serial buffer was empty. In this situation, the method would busy-wait and hog
the processor, never giving the serial buffer a chance to receive incoming
messages.

Solved this issue by introducing a small (1 ms) delay within the busy-wait
loop.