Reduced obscenely high baud rate.
Earlier, I believed the reason that the project was experiencing lag was because of a low baud rate, and I jacked it up to a very high level. It turns out that the baud rate was not causing the problem, and the extrememly high setting is actually now causing other issues on certain computers.

Reducing the baud rate to the standard 9600 Hz.